Exact random functions

+++ internal/models/player.go
@@ -1,8 +1,9 @@
 package models
 
 type Player struct {
-	ShContrib float64
-	TkContrib float64
-	PsContrib float64
-	Ag        float64
+	ShContrib int
+	TkContrib int
+	PsContrib int
+
+	Ag int
 }
